use bigDecimal, LocalDateTime accpet param,

+++ mybatis/src/main/java/mybatis/bean/Score.java
@@ -3,6 +3,8 @@ package mybatis.bean;
 import lombok.Data;
 import lombok.experimental.Accessors;
 
+import java.math.BigDecimal;
+
 /**
  * Create by peng.qd on 2019/2/24 0024.
  */
@@ -15,6 +17,6 @@ public class Score {
     //course id
     private String cId;
 
-    private Double score;
+    private BigDecimal score;
 
 }
